% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy.trimfill.R
\name{trimfill.rma}
\alias{trimfill.rma}
\title{Trim&fill method for univariate meta analysis}
\usage{
\method{trimfill}{rma}(
  x,
  side,
  estimator = "L0",
  maxiter = 100,
  method.trim = NULL,
  method.fill = NULL,
  verbose = FALSE,
  ilim
)
}
\arguments{
\item{x}{an object of class "rma.uni".}

\item{side}{the same as in metafor::trimfill}

\item{estimator}{the same as in metafor::trimfill}

\item{maxiter}{the same as in metafor::trimfill}

\item{method.trim}{the model used in rma.uni() for estimating the center when trimming studies, default is x$method}

\item{method.fill}{the model used in rma.uni() for estimating the center after filling studies, default is x$method}

\item{verbose}{the same as in metafor::trimfill}

\item{ilim}{limits for the imputed values as in metafor::trimfill. If unspecified, no limits are used.}
}
\value{
the same as in metafor::trimfill
}
\description{
Modified metafor::trimfill.rma.uni to avoid the invalid sqrt in k0 calculation when estimator == "Q0"
}
\details{
It is recommend using fixed-effects for method.trim and random-effects for method.fill when heterogeneity exists.
}
\author{
Chongliang Luo, Yong Chen
}
